requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
mediapipe>=0.10.9
insightface>=0.7.3
deep-sort-realtime>=1.3.2
opencv-python-headless>=4.9.0
ffmpeg-python>=0.2.0
onnxruntime-gpu>=1.17.0
PyQt6>=6.6.0
numpy>=1.24.0
scipy>=1.11.0
rich>=13.7.0
pytest>=8.0.0
Pillow>=10.2.0
scikit-image>=0.22.0
